+++ secantTests/BalanceBreakdownTests/BalanceBreakdownTests.swift
@@ -88,7 +88,7 @@ class BalanceBreakdownTests: XCTestCase {
                 isShieldingFunds: false,
                 pendingTransactions: .zero,
                 shieldedBalance: Balance.zero,
-                synchronizerStatusSnapshot: .initial,
+                syncProgressState: .initial,
                 transparentBalance: Balance(
                     WalletBalance(
                         verified: Zatoshi(1_000_000),
@@ -113,7 +113,7 @@ class BalanceBreakdownTests: XCTestCase {
                 isShieldingFunds: true,
                 pendingTransactions: .zero,
                 shieldedBalance: Balance.zero,
-                synchronizerStatusSnapshot: .initial,
+                syncProgressState: .initial,
                 transparentBalance: Balance(
                     WalletBalance(
                         verified: Zatoshi(1_000_000),
@@ -130,130 +130,30 @@ class BalanceBreakdownTests: XCTestCase {
         XCTAssertTrue(store.state.isShieldingButtonDisabled)
     }
     
-    func testSyncingData() async throws {
-        let store = TestStore(
-            initialState: BalanceBreakdownReducer.State(
-                autoShieldingThreshold: Zatoshi(1_000_000),
-                changePending: .zero,
-                isShieldingFunds: true,
-                pendingTransactions: .zero,
-                shieldedBalance: Balance.zero,
-                synchronizerStatusSnapshot: .snapshotFor(state: .syncing(0.513)),
-                transparentBalance: Balance(
-                    WalletBalance(
-                        verified: Zatoshi(1_000_000),
-                        total: Zatoshi(1_000_000)
-                    )
-                )
-            )
-        ) {
-            BalanceBreakdownReducer(networkType: .testnet)
-        }
-
-        XCTAssertTrue(store.state.isSyncing)
-        XCTAssertEqual(store.state.syncingPercentage, 0.513 * 0.999)
-    }
-    
-    func testlastKnownSyncingPercentage_Zero() async throws {
-        let store = TestStore(
-            initialState: BalanceBreakdownReducer.State(
-                autoShieldingThreshold: Zatoshi(1_000_000),
-                changePending: .zero,
-                isShieldingFunds: true,
-                pendingTransactions: .zero,
-                shieldedBalance: Balance.zero,
-                synchronizerStatusSnapshot: .placeholder,
-                transparentBalance: Balance(
-                    WalletBalance(
-                        verified: Zatoshi(1_000_000),
-                        total: Zatoshi(1_000_000)
-                    )
-                )
-            )
-        ) {
-            BalanceBreakdownReducer(networkType: .testnet)
-        }
-
-        XCTAssertEqual(store.state.lastKnownSyncPercentage, 0)
-        XCTAssertEqual(store.state.syncingPercentage, 0)
-    }
-    
-    func testlastKnownSyncingPercentage_MoreThanZero() async throws {
-        let store = TestStore(
-            initialState: BalanceBreakdownReducer.State(
-                autoShieldingThreshold: Zatoshi(1_000_000),
-                changePending: .zero,
-                isShieldingFunds: true,
-                lastKnownSyncPercentage: 0.15,
-                pendingTransactions: .zero,
-                shieldedBalance: Balance.zero,
-                synchronizerStatusSnapshot: .placeholder,
-                transparentBalance: Balance(
-                    WalletBalance(
-                        verified: Zatoshi(1_000_000),
-                        total: Zatoshi(1_000_000)
-                    )
-                )
-            )
-        ) {
-            BalanceBreakdownReducer(networkType: .testnet)
-        }
-
-        XCTAssertEqual(store.state.lastKnownSyncPercentage, 0.15)
-        XCTAssertEqual(store.state.syncingPercentage, 0.15)
-    }
-    
-    func testlastKnownSyncingPercentage_FromSyncedState() async throws {
-        let store = TestStore(
-            initialState: BalanceBreakdownReducer.State(
-                autoShieldingThreshold: Zatoshi(1_000_000),
-                changePending: .zero,
-                isShieldingFunds: true,
-                lastKnownSyncPercentage: 0.15,
-                pendingTransactions: .zero,
-                shieldedBalance: Balance.zero,
-                synchronizerStatusSnapshot: .snapshotFor(state: .syncing(0.513)),
-                transparentBalance: .zero
-            )
-        ) {
-            BalanceBreakdownReducer(networkType: .testnet)
-        }
-
-        var syncState: SynchronizerState = .zero
-        syncState.syncStatus = .upToDate
-        let snapshot = SyncStatusSnapshot.snapshotFor(state: syncState.syncStatus)
+    func testRestoreWalletSubscription() async throws {
+        var initialState = BalanceBreakdownReducer.State.initial
+        initialState.isRestoringWallet = false
         
-        await store.send(.synchronizerStateChanged(syncState)) { state in
-            state.synchronizerStatusSnapshot = snapshot
-            state.syncStatusMessage = "Synced"
-            state.lastKnownSyncPercentage = 1.0
-        }
-    }
-    
-    func testlastKnownSyncingPercentage_FromSyncingState() async throws {
         let store = TestStore(
-            initialState: BalanceBreakdownReducer.State(
-                autoShieldingThreshold: Zatoshi(1_000_000),
-                changePending: .zero,
-                isShieldingFunds: true,
-                lastKnownSyncPercentage: 0.15,
-                pendingTransactions: .zero,
-                shieldedBalance: Balance.zero,
-                synchronizerStatusSnapshot: .snapshotFor(state: .syncing(0.513)),
-                transparentBalance: .zero
-            )
+            initialState: initialState
         ) {
             BalanceBreakdownReducer(networkType: .testnet)
         }
 
-        var syncState: SynchronizerState = .zero
-        syncState.syncStatus = .syncing(0.545)
-        let snapshot = SyncStatusSnapshot.snapshotFor(state: syncState.syncStatus)
-        
-        await store.send(.synchronizerStateChanged(syncState)) { state in
-            state.synchronizerStatusSnapshot = snapshot
-            state.syncStatusMessage = "Syncing"
-            state.lastKnownSyncPercentage = 0.545
+        store.dependencies.restoreWalletStorage = .noOp
+        store.dependencies.restoreWalletStorage.value = {
+            AsyncStream { continuation in
+                continuation.yield(true)
+                continuation.finish()
+            }
         }
+        
+        await store.send(.restoreWalletTask)
+        
+        await store.receive(.restoreWalletValue(true)) { state in
+            state.isRestoringWallet = true
+        }
+        
+        await store.finish()
     }
 }