+++ secantTests/RootTests/AppInitializationTests.swift
@@ -54,6 +54,7 @@ class AppInitializationTests: XCTestCase {
         store.dependencies.sdkSynchronizer = .noOp
         store.dependencies.crashReporter = .noOp
         store.dependencies.numberFormatter = .noOp
+        store.dependencies.restoreWalletStorage = .noOp
 
         // Root of the test, the app finished the launch process and triggers the checks and initializations.
         await store.send(.initialization(.appDelegate(.didFinishLaunching)))
@@ -127,6 +128,7 @@ class AppInitializationTests: XCTestCase {
         store.dependencies.sdkSynchronizer = .noOp
         store.dependencies.crashReporter = .noOp
         store.dependencies.numberFormatter = .noOp
+        store.dependencies.restoreWalletStorage = .noOp
 
         // Root of the test, the app finished the launch process and triggers the checks and initializations.
         await store.send(.initialization(.appDelegate(.didFinishLaunching)))
@@ -184,6 +186,7 @@ class AppInitializationTests: XCTestCase {
         store.dependencies.mainQueue = .immediate
         store.dependencies.walletConfigProvider = .noOp
         store.dependencies.crashReporter = .noOp
+        store.dependencies.restoreWalletStorage = .noOp
 
         // Root of the test, the app finished the launch process and triggers the checks and initializations.
         await store.send(.initialization(.appDelegate(.didFinishLaunching)))
@@ -215,6 +218,7 @@ class AppInitializationTests: XCTestCase {
         store.dependencies.walletStorage = .noOp
         store.dependencies.walletConfigProvider = .noOp
         store.dependencies.crashReporter = .noOp
+        store.dependencies.restoreWalletStorage = .noOp
 
         // Root of the test, the app finished the launch process and triggers the checks and initializations.
         await store.send(.initialization(.appDelegate(.didFinishLaunching)))